requirements.txt
django
openai
djangorestframework